requirements.txt
gradio==4.44.1
PyMuPDF==1.23.0
requests==2.31.0
duckduckgo-search>=6.0.0
google-generativeai>=0.7.1